CO-0007: Refactor and project organization
