+++ DubDubGrub/Extensions/Color+Extension.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 extension Color {
-    
-    static let brandPrimary = Color("brandPrimary")
+    static let brandPrimary     = Color("brandPrimary")
+    static let brandSecondary   = Color(.systemPink)
+    static let grubRed          = Color(.systemPink)
 }
