+++ DubDubGrub/Extensions/View+Extension.swift
@@ -1,0 +1,14 @@
+//
+//  View+Extension.swift
+//  DubDubGrub
+//
+//  Created by Oscar Cristaldo on 2022-08-23.
+//
+
+import SwiftUI
+
+extension View {
+    func profileNameStyle() -> some View {
+        self.modifier(ProfileNameText())
+    }
+}
