--- DubDubGrub/LocationListView.swift
@@ -1,71 +1,0 @@
-//
-//  LocationListView.swift
-//  DubDubGrub
-//
-//  Created by Oscar Cristaldo on 2022-08-09.
-//
-
-import SwiftUI
-
-struct LocationListView: View {
-    var body: some View {
-        NavigationView {
-            List {
-                ForEach(0..<10) { item in
-                    NavigationLink(destination: LocationDetailView()) {
-                        LocationCell()
-                    }
-                }
-            }.listStyle(PlainListStyle())
-            .navigationTitle("Grub Spots")
-        }
-        .navigationViewStyle(.stack)
-    }
-}
-
-struct LocationListView_Previews: PreviewProvider {
-    static var previews: some View {
-        LocationListView()
-    }
-}
-
-struct AvatarView: View {
-    
-    var size: CGFloat
-    
-    var body: some View {
-        Image("default-avatar")
-            .resizable()
-            .scaledToFit()
-            .frame(width: size, height: size)
-            .clipShape(Circle())
-    }
-}
-
-struct LocationCell: View {
-    var body: some View {
-        HStack {
-            Image("default-square-asset")
-                .resizable()
-                .scaledToFit()
-                .frame(width: 80, height: 80)
-                .clipShape(Circle())
-                .padding(.vertical, 8)
-            
-            VStack(alignment: .leading) {
-                Text("Test Location Name")
-                    .font(.title2)
-                    .fontWeight(.semibold)
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.75)
-                HStack {
-                    AvatarView(size: 35)
-                    AvatarView(size: 35)
-                    AvatarView(size: 35)
-                    AvatarView(size: 35)
-                }
-            }
-            .padding(.leading)
-        }
-    }
-}
