--- DubDubGrub/ProfileView.swift
@@ -1,109 +1,0 @@
-//
-//  ProfileView.swift
-//  DubDubGrub
-//
-//  Created by Oscar Cristaldo on 2022-08-09.
-//
-
-import SwiftUI
-
-struct ProfileView: View {
-    
-    @State
-    private var firstName       = ""
-    
-    @State
-    private var lastName        = ""
-    
-    @State
-    private var companyName     = ""
-    
-    @State
-    private var bio             = ""
-    
-    var body: some View {
-        VStack {
-            ZStack {
-                Color(uiColor: .secondarySystemBackground)
-                    .frame(height: 130)
-                    .cornerRadius(12)
-                    .padding(.horizontal)
-                
-                HStack(spacing: 16) {
-                    ZStack {
-                        AvatarView(size: 84)
-                        
-                        Image(systemName: "square.and.pencil")
-                            .resizable()
-                            .scaledToFit()
-                            .frame(width: 14, height: 14)
-                            .foregroundColor(.white)
-                            .offset(y: 30)
-                    }
-                    .padding(.leading, 12)
-                    
-                    VStack(spacing: 1) {
-                        TextField("First Name", text: $firstName)
-                            .font(.system(size: 32, weight: .bold))
-                            .lineLimit(1)
-                            .minimumScaleFactor(0.75)
-                        
-                        TextField("Last Name", text: $lastName)
-                            .font(.system(size: 32, weight: .bold))
-                            .lineLimit(1)
-                            .minimumScaleFactor(0.75)
-                        
-                        TextField("Company Name", text: $companyName)
-                    }
-                    .padding(.trailing, 16)
-                }
-                .padding()
-            }
-            
-            VStack(alignment: .leading, spacing: 8) {
-                Text("Bio: ")
-                    .font(.callout)
-                    .foregroundColor(.secondary)
-                +
-                Text("\(100 - bio.count)")
-                    .bold()
-                    .font(.callout)
-                    .foregroundColor(bio.count <= 100 ? Color("AccentColor") : Color(uiColor: .systemPink))
-                +
-                Text(" Characters Remain")
-                    .font(.callout)
-                    .foregroundColor(.secondary)
-                
-                TextEditor(text: $bio)
-                    .frame(height: 100)
-                    .overlay(RoundedRectangle(cornerRadius: 8)
-                        .stroke(Color.secondary, lineWidth: 1))
-                
-            }
-            .padding(.horizontal, 20)
-            
-            Spacer()
-            
-            Button {
-                
-            } label: {
-                Text("Create Profile")
-                    .bold()
-                    .frame(width: 280, height: 44)
-                    .background(Color("AccentColor"))
-                    .foregroundColor(.white)
-                    .cornerRadius(8)
-            }
-        }
-        .navigationTitle("Profile")
-    }
-}
-
-struct ProfileView_Previews: PreviewProvider {
-    static var previews: some View {
-        NavigationView {
-            ProfileView()
-        }
-        
-    }
-}
