+++ DubDubGrub/Views/CustomModifiers.swift
@@ -1,0 +1,17 @@
+//
+//  CustomModifiers.swift
+//  DubDubGrub
+//
+//  Created by Oscar Cristaldo on 2022-08-23.
+//
+
+import SwiftUI
+
+struct ProfileNameText: ViewModifier {
+    func body(content: Content) -> some View {
+        content
+            .font(.system(size: 32, weight: .bold))
+            .lineLimit(1)
+            .minimumScaleFactor(0.75)
+    }
+}
