+++ DubDubGrub/Views/CustomViews/DDGButton.swift
@@ -1,0 +1,28 @@
+//
+//  DDGButton.swift
+//  DubDubGrub
+//
+//  Created by Oscar Cristaldo on 2022-08-23.
+//
+
+import SwiftUI
+
+struct DDGButton: View {
+    
+    var title: String
+    
+    var body: some View {
+        Text(title)
+            .bold()
+            .frame(width: 280, height: 44)
+            .background(Color("AccentColor"))
+            .foregroundColor(.white)
+            .cornerRadius(8)
+    }
+}
+
+struct DDGButton_Previews: PreviewProvider {
+    static var previews: some View {
+        DDGButton(title: "Test Button")
+    }
+}
