+++ DubDubGrub/Views/Screens/AppTabView.swift
@@ -8,11 +8,16 @@
 import SwiftUI
 
 struct AppTabView: View {
+    
+    init() {
+        UITabBar.appearance().backgroundColor = UIColor.secondarySystemBackground
+    }
+    
     var body: some View {
         TabView {
             LocationMapView().tabItem{ Label("Map", systemImage: "map") }
             LocationListView().tabItem{ Label("Locations", systemImage: "building") }
-            ProfileView().tabItem{ Label("Profile", systemImage: "person") }
+            NavigationView { ProfileView() }.tabItem{ Label("Profile", systemImage: "person") }
         }.tint(Color("AccentColor"))
     }
 }
