+++ DubDubGrub/Views/Screens/LocationListView.swift
@@ -1,0 +1,30 @@
+//
+//  LocationListView.swift
+//  DubDubGrub
+//
+//  Created by Oscar Cristaldo on 2022-08-09.
+//
+
+import SwiftUI
+
+struct LocationListView: View {
+    var body: some View {
+        NavigationView {
+            List {
+                ForEach(0..<10) { item in
+                    NavigationLink(destination: LocationDetailView()) {
+                        LocationCell()
+                    }
+                }
+            }.listStyle(PlainListStyle())
+            .navigationTitle("Grub Spots")
+        }
+        .navigationViewStyle(.stack)
+    }
+}
+
+struct LocationListView_Previews: PreviewProvider {
+    static var previews: some View {
+        LocationListView()
+    }
+}
