+++ DubDubGrub/Views/Screens/ProfileView.swift
@@ -1,0 +1,109 @@
+//
+//  ProfileView.swift
+//  DubDubGrub
+//
+//  Created by Oscar Cristaldo on 2022-08-09.
+//
+
+import SwiftUI
+
+struct ProfileView: View {
+    
+    @State private var firstName    = ""
+    @State private var lastName     = ""
+    @State private var companyName  = ""
+    @State private var bio          = ""
+    
+    var body: some View {
+        VStack {
+            ZStack {
+                NameBackgroundView()
+                
+                HStack(spacing: 16) {
+                    ZStack {
+                        AvatarView(size: 84)
+                        EditImageView()
+                    }
+                    .padding(.leading, 12)
+                    
+                    VStack(spacing: 1) {
+                        TextField("First Name", text: $firstName).profileNameStyle()
+                        TextField("Last Name", text: $lastName).profileNameStyle()
+                        TextField("Company Name", text: $companyName)
+                    }
+                    .padding(.trailing, 16)
+                }
+                .padding()
+            }
+            
+            VStack(alignment: .leading, spacing: 8) {
+                CharacterRemainView(currentCount: bio.count)
+                
+                TextEditor(text: $bio)
+                    .frame(height: 100)
+                    .overlay(RoundedRectangle(cornerRadius: 8)
+                        .stroke(Color.secondary, lineWidth: 1))
+                
+            }
+            .padding(.horizontal, 20)
+            
+            Spacer()
+            
+            Button {
+                
+            } label: {
+                DDGButton(title: "Create Profile")
+            }
+        }
+        .navigationTitle("Profile")
+    }
+}
+
+struct ProfileView_Previews: PreviewProvider {
+    static var previews: some View {
+        NavigationView {
+            ProfileView()
+        }
+        
+    }
+}
+
+struct NameBackgroundView: View {
+    var body: some View {
+        Color(uiColor: .secondarySystemBackground)
+            .frame(height: 130)
+            .cornerRadius(12)
+            .padding(.horizontal)
+    }
+}
+
+struct EditImageView: View {
+    var body: some View {
+        Image(systemName: "square.and.pencil")
+            .resizable()
+            .scaledToFit()
+            .frame(width: 14, height: 14)
+            .foregroundColor(.white)
+            .offset(y: 30)
+    }
+}
+
+struct CharacterRemainView: View {
+    
+    var currentCount: Int
+    
+    var body: some View {
+        Text("Bio: ")
+            .font(.callout)
+            .foregroundColor(.secondary)
+        +
+        Text("\(100 - currentCount)")
+            .bold()
+            .font(.callout)
+            .foregroundColor(currentCount <= 100 ? Color("AccentColor") : Color(uiColor: .systemPink))
+        +
+        Text(" Characters Remain")
+            .font(.callout)
+            .foregroundColor(.secondary)
+    }
+}
